pyproject.toml
[project]
name = "data-validator-dblink"
version = "0.1.0"
description = "Oracle Migration Data Validator - Configuration-driven data validation for Oracle database migrations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "oracledb>=2.0.1",
    "pydantic>=2.0.0",
    "email-validator>=2.1.0",
    "aiofiles>=23.2.1",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
    "multiprocessing-logging>=0.3.4"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]
